requirements.txt
asgiref==3.7.2
astroid==2.15.6
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-phonenumber-field==7.1.0
django-stubs==4.2.3
django-stubs-ext==4.2.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
flake8==6.1.0
flake8-django==1.4
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.17
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
